Add files via upload
Nowadays Tailwind CSS is going more popular with the use of any javascript frameworks like Reactjs and Vuejs. But here we simply used jquery click function and also used Tailwind CSS component and Utility classes.

This is a responsive element for mobile and tablet views. Here we mostly used Utility classes that are already given in this framework. Also for mobile toggle clickable, we use jquery CDN and click function with the help of the toggle method. To make responsive we didn’t use any single line of CSS or media quires.

Read more:-
http://hots.link/Pvi